Feature security

+++ user-service/src/main/java/ru/gnidenko/userservice/controller/UserController.java
@@ -1,23 +1,24 @@
 package ru.gnidenko.userservice.controller;
 
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ru.gnidenko.userservice.dto.CreateRequestUserDto;
 import ru.gnidenko.userservice.dto.UserDto;
 import ru.gnidenko.userservice.exception.DataValidationException;
 import ru.gnidenko.userservice.service.UserService;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Set;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -25,45 +26,56 @@ import java.util.Objects;
 public class UserController {
     private final UserService userService;
 
-    @PostMapping
-    public UserDto addUser(@Valid @RequestBody CreateRequestUserDto userDto, HttpServletResponse response) {
-        checkBodyIsNotNull(userDto);
-        response.setStatus(HttpServletResponse.SC_CREATED);
-        return userService.addUser(userDto);
-    }
-
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Long id, HttpServletResponse response) {
+        if (!getSetRolesFromClaims().contains("ADMIN")) {
+            throw new BadCredentialsException("You are not allowed");
+        }
         checkBodyIsNotNull(id);
         response.setStatus(HttpServletResponse.SC_OK);
         userService.deleteUser(id);
     }
 
-    @GetMapping("/{id}")
-    public UserDto getUser(@PathVariable Long id, HttpServletResponse response) {
-        checkBodyIsNotNull(id);
+    @GetMapping("/me")
+    public UserDto getUser(HttpServletResponse response) {
+        Long currentId = getUserIdFromClaims();
         response.setStatus(HttpServletResponse.SC_OK);
-        return userService.findUser(id);
+        return userService.findUser(currentId);
     }
 
     @GetMapping
     public List<UserDto> getAllUsers(HttpServletResponse response) {
+        if (!getSetRolesFromClaims().contains("ADMIN")) {
+            throw new BadCredentialsException("You are not allowed");
+        }
         response.setStatus(HttpServletResponse.SC_OK);
         return userService.findAllUsers();
     }
 
-    @PutMapping("/{id}")
-    public UserDto updateUser(@RequestBody UserDto userDto, @PathVariable Long id, HttpServletResponse response) {
+    @PutMapping
+    public UserDto updateUser(@RequestBody UserDto userDto, HttpServletResponse response) {
         checkBodyIsNotNull(userDto);
-        checkBodyIsNotNull(id);
+        Long currentId = getUserIdFromClaims();
         response.setStatus(HttpServletResponse.SC_OK);
-        return userService.updateUser(userDto, id);
+        return userService.updateUser(userDto, currentId);
     }
 
-    private <T> void checkBodyIsNotNull(T body){
-        if (Objects.isNull(body)){
+    private <T> void checkBodyIsNotNull(T body) {
+        if (Objects.isNull(body)) {
             throw new DataValidationException("body is null");
         }
+    }
+
+    private Long getUserIdFromClaims() {
+        return getJwtFromPrincipal().getClaim("userId");
+    }
+
+    private List<String> getSetRolesFromClaims() {
+        return getJwtFromPrincipal().getClaim("roles");
+    }
+
+    private Jwt getJwtFromPrincipal() {
+        return (Jwt) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 
 }
